Migrate DaggerSuperficialValidation related tests to XProcessing.
This CL firsts migrates DaggerSuperficialValidation to use `toStableString()` for reporting elements, types, etc. This is required to ensure that the error messages match between javac and ksp.

In addition, given that we are trying to report string representations of error/invalid types, XProcessing may throw a TypeNotPresentException in some cases. Thus, we wrap each `toStableString()` implementation in try-catch and report TypeNotPresentException#typeName() when this error is thrown.

RELNOTES=N/A
PiperOrigin-RevId: 479591941