Use XMethodElement#getJvmName() explicitly in places where we call the method.
This CL _should not_ be a functional change because `XElements#getSimpleName(XElement)` has always returned the jvm name for methods. The purpose of this CL is 1) to add more documentation on why we've chosen this, and 2) to explicitly use getJvmName() in places where we actually need to call the method since this is technically more correct (and we want things to continue working if we ever fix `getSimpleName()`).

RELNOTES=N/A
PiperOrigin-RevId: 492027457