Remove unused Dagger test sources.
These sources were copied over to the `functional/membersinject/subpackage` package as part of fixing the `glob(**/*)` in our tests. However, on closer inspection, the component that adds `AModule.java` doesn't actually use it in this case -- i.e. it will get pruned from the graph since none of the entry points transitively depend on it. Thus, I'm just deleting it in this CL to clean up our tests.

RELNOTES=N/A
PiperOrigin-RevId: 495357856