Refactor MembersWithSameNameTest in preparation for kotlin sources.
This CL removes the glob(**) by adding a build file to the subpackage of this test (note: I had to move some sources to avoid build cycles). This change will be useful when converting the test to kotlin sources because allowing the sources to live in separate build units allows us to replace the package-private visibility with internal visibility in the test. I've also done some renaming and added some comments to make things a little easier to parse since it took me a bit of time to parse what this test was trying to cover.

RELNOTES=N/A
PiperOrigin-RevId: 497009423