Add suppress warnings to allow Dagger's generated java sources to reference kotlin internal types.
This suppression is needed for Dagger's generated code which doesn't properly account for Kotlin `internal` visibility. In particular, the bytecode for `internal` is just `public` visibility so Dagger's generated code treats the type as if it were public.

Eventually, we'll fix things so that this suppression isn't needed, but that will require a bit of work on the generated sources and will also cause potential version skew issues between dependencies that use different versions of Dagger. Thus, for now we're just suppressing these warnings rather than fixing them.

RELNOTES=N/A
PiperOrigin-RevId: 499252778