Cleanup dagger/functional/nullables tests.
This CL moves all classes related to NullabilityTest into the test itself so that all necessary dependencies are encapsulated within the test. This helps readability when there are multiple tests in package so that you don't have to figure out which sources are used by which test.

RELNOTES=N/A
PiperOrigin-RevId: 506516080