[Clean-up]: Remove unnecessary usage of Hilt's ComponentDescriptors.
Hilt's `ComponentDescriptor` shouldn't be used outside of the root processors because they rely on having the full set of components + builders in the transitive classpath, which may not be true at any place other than the root.

In this case `DefineComponentValidationProcessor` doesn't actually need the extra information provided by the `ComponentDescriptor` so I've changed it to just use the `DefineComponentMetadata` instead, which just contains the information contained on the `@DefineComponent` element itself.

RELNOTES=N/A
PiperOrigin-RevId: 513065950