Migrate DeclarationVarianceTest to a compiler tests.
This change is in preparation for an upcoming feature where Dagger will ignore the variance in Key types. Once this feature is implemented, the tests in DeclarationVarainceTest will no longer succeed since the keys will be considered duplicate bindings. This CL migrates the tests to compiler tests so that we can implement the feature and also test the error messages associated with it.

RELNOTES=N/A
PiperOrigin-RevId: 535291620