Cleanup usages of `Key.Builder`, use `KeyFactory` instead.
In a follow-up CL we will need to use the `CompilerOptions` when creating keys, so this CL replaces usages of `Key.Builder` with `KeyFactory` where we can easily get access to the `CompilerOptions`.

In addition, I've removed some unnecessary parts of `KeyFactoryTest` that was testing more of the implementation details of `KeyFactory` and relied heavily on `Key.Builder`.

RELNOTES=N/A
PiperOrigin-RevId: 535366633