Add script to automate GitHub Actions cache clean up on each run.
This CL fixes an issues in our GitHub Actions where we are consistently going over our cache limit. Typically, we would need to clean out the cache manually at https://github.com/google/dagger/actions/caches. This script automates that process by deleting caches each run.

  - For a PR that is already closed: deletes all corresponding caches.
  - For duplicate caches with the same Git reference and cache key: deletes the older cache.

The script should be idempotent, so I decided to run the script both at the beginning and end of each workflow. This ensure that the caches are cleaned up after a workflow ends, but also that we don't start accumulating unnecessary caches when a workflow fails before the script can be run.

RELNOTES=N/A
PiperOrigin-RevId: 551349636