Add @Nullable annotation in error message when formatting methods with nullable types.
This CL adds an @Nullable annotation in the error message when formatting binding methods with nullable types like `Foo?`. This behavior matches KAPT's generated java stubs which will add an `@Nullable` annotation to the stub's method when the return type of the method is a nullable type.

RELNOTES=N/A
PiperOrigin-RevId: 554830004